Refactor utilities into focused modules
