+++ README.md
@@ -23,6 +23,8 @@ A small Tkinter-based interface for sending prompts to the [`aider`](https://git
 
 - Add tests for functionality changes to verify new behavior.
 - Break up and modularize code where possible to reduce merge conflicts.
+- Utility helpers live in focused modules (e.g., `utils.text`, `utils.config`)
+  so changes stay scoped to their domain.
 
 ## Author
 Ben Arnao
